Additional postgres handling details
The database is identified as "postgres9" rather than "pgsql"
to align with ADOdb naming

Selecting a database or schema uses an adapter to paper
over different semantics

Connecting to a postgres database assumes that the database
name corresponds to the user name. This allows existing "database name"
fields to store schema names and avoids adding any additional
configuration option(s) that would be postgres-specific.

Conflicts:
	common/SQLManager.php
	common/sql/Lib.php