SPINS date handling
When the offset from ISO week numbering to SPINS week numbering
if negative, the default week for the submit task ends up
multiple weeks in the past. All data still gets sent eventually
but not by period deadlines. After figuring out the week
numbering scheme and a pair of start/end dates that match a
particular week number it now walks forward in seven day
increments so that the default week is the most recent finished
week.