Use NamedSession in Office
This change is to better facilitate side-by-side installs.
FannieAPI caches information about where classes are defined
in $_SESSION. If multiple copies of Office are installed on
the same server, session collisions mean it's possible for
the first copy of Office to load a class from within the second
copy of Office or vice versa. If the side-by-side installs are
different versions this can cause lots of chaos

By using NamedSession and the filesystem path where Office is
installed, each copy of Office attaches a unique prefix to its
session variables. This is purely a usability concern. These
session values should not be considered secure from other PHP
code on the same server. The whole session is still accessible.